@@ -1,457 +1,553 @@
 <template>
   <div id="gallery-view">
     <div class="view-title">
-      相册 - {{ filterList.length }} <i class="el-icon-caret-bottom" @click="toggleHandleBar" :class="{'active': handleBarActive}"></i>
+      {{ $T('GALLERY') }} - {{ filterList.length }}
+      <el-icon
+        style="margin-left: 4px"
+        class="cursor-pointer"
+        @click="toggleHandleBar"
+      >
+        <CaretBottom v-show="!handleBarActive" />
+        <CaretTop v-show="handleBarActive" />
+      </el-icon>
     </div>
     <transition name="el-zoom-in-top">
       <el-row v-show="handleBarActive">
-        <el-col :span="20" :offset="2">
-          <el-row class="handle-bar" :gutter="16">
+        <el-col
+          :span="20"
+          :offset="2"
+        >
+          <el-row
+            class="handle-bar"
+            :gutter="16"
+          >
             <el-col :span="12">
               <el-select
-                v-model="choosedPicBed"
+                v-model="selectedPicBed"
                 multiple
                 collapse-tags
-                size="mini"
+                size="small"
                 style="width: 100%"
-                placeholder="请选择显示的图床">
+                :placeholder="$T('CHOOSE_SHOWED_PICBED')"
+              >
                 <el-option
                   v-for="item in picBed"
                   :key="item.type"
                   :label="item.name"
-                  :value="item.type">
-                </el-option>
+                  :value="item.type"
+                />
               </el-select>
             </el-col>
             <el-col :span="12">
               <el-select
                 v-model="pasteStyle"
-                size="mini"
+                size="small"
                 style="width: 100%"
+                :placeholder="$T('CHOOSE_PASTE_FORMAT')"
                 @change="handlePasteStyleChange"
-                placeholder="请选择粘贴的格式">
+              >
                 <el-option
                   v-for="(value, key) in pasteStyleMap"
                   :key="key"
                   :label="key"
-                  :value="value">
-                </el-option>
+                  :value="value"
+                />
               </el-select>
             </el-col>
           </el-row>
-          <el-row class="handle-bar" :gutter="16">
+          <el-row
+            class="handle-bar"
+            :gutter="16"
+          >
             <el-col :span="12">
               <el-input
-                placeholder="搜索"
-                size="mini"
-                v-model="searchText">
-                <i slot="suffix" class="el-input__icon el-icon-close" v-if="searchText" @click="cleanSearch" style="cursor: pointer"></i>
+                v-model="searchText"
+                :placeholder="$T('SEARCH')"
+                size="small"
+              >
+                <template #suffix>
+                  <el-icon
+                    class="el-input__icon"
+                    style="cursor: pointer;"
+                    @click="cleanSearch"
+                  >
+                    <close />
+                  </el-icon>
+                </template>
               </el-input>
             </el-col>
-            <el-col :span="4">
-              <div class="item-base copy round" :class="{ active: isMultiple(choosedList)}" @click="multiCopy">
-                复制
-              </div>
-            </el-col>
-            <el-col :span="4">
-              <div class="item-base delete round" :class="{ active: isMultiple(choosedList)}" @click="multiRemove">
-                删除
-              </div>
-            </el-col>
-            <el-col :span="4">
-              <div class="item-base all-pick round" :class="{ active: filterList.length > 0}" @click="toggleSelectAll">
-                {{ isAllSelected ? '取消' : '全选' }}
-              </div>
-            </el-col>
+            <GalleryToolbar
+              :selected-list="selectedList"
+              :filter-list="filterList"
+              :is-all-selected="isAllSelected"
+              @multi-copy="multiCopy"
+              @multi-remove="multiRemove"
+              @toggle-select-all="toggleSelectAll"
+              @select-more="selectMore"
+            />
           </el-row>
         </el-col>
       </el-row>
     </transition>
-    <el-row class="gallery-list" :class="{ small: handleBarActive }">
-      <el-col :span="20" :offset="2">
+    <el-row
+      class="gallery-list"
+      :class="{ small: handleBarActive }"
+    >
+      <el-col
+        :span="20"
+        :offset="2"
+      >
         <el-row :gutter="16">
-          <gallerys
-            :images="filterList"
-            :index="idx"
-            @close="handleClose"
-            :options="options"
-          ></gallerys>
-          <el-col :span="6" v-for="(item, index) in filterList" :key="item.id" class="gallery-list__img">
+          <photo-slider
+            :items="filterList"
+            :visible="gallerySliderControl.visible"
+            :index="gallerySliderControl.index"
+            :should-transition="true"
+            @change-index="zoomImage"
+            @click-mask="handleClose"
+            @close-modal="handleClose"
+          />
+          <el-col
+            v-for="(item, index) in filterList"
+            :key="item.id"
+            :span="6"
+            class="gallery-list__img"
+          >
             <div
               class="gallery-list__item"
               @click="zoomImage(index)"
             >
-              <img v-lazy="item.imgUrl" class="gallery-list__item-img">
+              <img
+                v-lazy="item.imgUrl"
+                class="gallery-list__item-img"
+              >
             </div>
-            <div class="gallery-list__tool-panel">
-              <i class="el-icon-document" @click="copy(item)"></i>
-              <i class="el-icon-edit-outline" @click="openDialog(item)"></i>
-              <i class="el-icon-delete" @click="remove(item.id)"></i>
-              <el-checkbox v-model="choosedList[item.id]" class="pull-right" @change="(val) => handleChooseImage(val, index)"></el-checkbox>
+            <div
+              class="gallery-list__file-name"
+              :title="item.fileName"
+            >
+              {{ item.fileName }}
             </div>
+            <el-row
+              class="gallery-list__tool-panel"
+              justify="space-between"
+              align="middle"
+            >
+              <el-row>
+                <el-icon
+                  class="cursor-pointer document"
+                  @click="copy(item)"
+                >
+                  <Document />
+                </el-icon>
+                <el-icon
+                  class="cursor-pointer edit"
+                  @click="openDialog(item)"
+                >
+                  <Edit />
+                </el-icon>
+                <el-icon
+                  class="cursor-pointer delete"
+                  @click="remove(item.id)"
+                >
+                  <Delete />
+                </el-icon>
+              </el-row>
+              <el-checkbox
+                v-model="selectedList[item.id ? item.id : '']"
+                @change="(val) => handleChooseImage(val, index)"
+              />
+            </el-row>
           </el-col>
         </el-row>
       </el-col>
     </el-row>
     <el-dialog
-      :visible.sync="dialogVisible"
-      title="修改图片URL"
+      v-model="dialogVisible"
+      :title="$T('CHANGE_IMAGE_URL')"
       width="500px"
       :modal-append-to-body="false"
     >
-      <el-input v-model="imgInfo.imgUrl"></el-input>
-      <span slot="footer" class="dialog-footer">
-        <el-button @click="dialogVisible = false">取 消</el-button>
-        <el-button type="primary" @click="confirmModify">确 定</el-button>
-      </span>
+      <el-input v-model="imgInfo.imgUrl" />
+      <template
+        #footer
+      >
+        <el-button @click="dialogVisible = false">
+          {{ $T('CANCEL') }}
+        </el-button>
+        <el-button
+          type="primary"
+          @click="confirmModify"
+        >
+          {{ $T('CONFIRM') }}
+        </el-button>
+      </template>
     </el-dialog>
   </div>
 </template>
-<script lang="ts">
-// @ts-ignore
-import gallerys from 'vue-gallery'
-import { Component, Vue, Watch } from 'vue-property-decorator'
-import { IResult } from '@picgo/store/dist/types'
-import { PASTE_TEXT } from '#/events/constants'
+<script lang="ts" setup>
+import type { IResult } from '@picgo/store/dist/types'
+import { PASTE_TEXT, GET_PICBEDS } from '#/events/constants'
+import { CheckboxValueType, ElMessageBox } from 'element-plus'
+import { Close, CaretBottom, Document, Edit, Delete, CaretTop } from '@element-plus/icons-vue'
 import {
   ipcRenderer,
   clipboard,
   IpcRendererEvent
 } from 'electron'
-@Component({
-  name: 'gallery',
-  components: {
-    gallerys
+import { computed, nextTick, onActivated, onBeforeUnmount, onBeforeMount, reactive, ref, watch } from 'vue'
+import { getConfig, saveConfig, sendRPC, sendToMain } from '@/utils/dataSender'
+import { onBeforeRouteUpdate } from 'vue-router'
+import { T as $T } from '@/i18n/index'
+import $$db from '@/utils/db'
+import GalleryToolbar from './components/gallery/GalleryToolbar.vue'
+import { IRPCActionType } from '~/universal/types/enum'
+import { getRawData } from '@/utils/common'
+const images = ref<ImgInfo[]>([])
+const dialogVisible = ref(false)
+const imgInfo = reactive({
+  id: '',
+  imgUrl: ''
+})
+const $confirm = ElMessageBox.confirm
+const selectedList: IObjT<boolean> = reactive({})
+const gallerySliderControl = reactive({
+  visible: false,
+  index: 0
+})
+const selectedPicBed = ref<string[]>([])
+const lastSelected = ref<number>(-1)
+const isShiftKeyPress = ref<boolean>(false)
+const searchText = ref<string>('')
+const handleBarActive = ref<boolean>(false)
+const pasteStyle = ref<string>('')
+const pasteStyleMap = {
+  Markdown: 'markdown',
+  HTML: 'HTML',
+  URL: 'URL',
+  UBB: 'UBB',
+  Custom: 'Custom'
+}
+const picBed = ref<IPicBedType[]>([])
+onBeforeRouteUpdate((to, from) => {
+  if (from.name === 'gallery') {
+    clearSelectedList()
+  }
+  if (to.name === 'gallery') {
+    updateGallery()
   }
 })
-export default class extends Vue {
-  images: ImgInfo[] = []
-  idx: null | number = null
-  options = {
-    titleProperty: 'fileName',
-    urlProperty: 'imgUrl',
-    closeOnSlideClick: true
-  }
 
-  dialogVisible = false
-  imgInfo = {
-    id: '',
-    imgUrl: ''
-  }
-
-  choosedList: IObjT<boolean> = {}
-  choosedPicBed: string[] = []
-  lastChoosed: number = -1
-  isShiftKeyPress: boolean = false
-  searchText = ''
-  handleBarActive = false
-  pasteStyle = ''
-  pasteStyleMap = {
-    Markdown: 'markdown',
-    HTML: 'HTML',
-    URL: 'URL',
-    UBB: 'UBB',
-    Custom: 'Custom'
-  }
-
-  picBed: IPicBedType[] = []
-  @Watch('$route')
-  handleRouteUpdate (to: any, from: any) {
-    console.log(to, from)
-    if (from.name === 'gallery') {
-      this.clearChoosedList()
-    }
-    if (to.name === 'gallery') {
-      this.updateGallery()
-    }
-  }
-
-  async created () {
-    ipcRenderer.on('updateGallery', () => {
-      this.$nextTick(async () => {
-        this.updateGallery()
-      })
+onBeforeMount(async () => {
+  ipcRenderer.on(IRPCActionType.UPDATE_GALLERY, () => {
+    nextTick(async () => {
+      updateGallery()
     })
-    ipcRenderer.send('getPicBeds')
-    ipcRenderer.on('getPicBeds', this.getPicBeds)
-    this.updateGallery()
-  }
+  })
+  sendToMain(GET_PICBEDS)
+  ipcRenderer.on(GET_PICBEDS, getPicBeds)
+  updateGallery()
 
-  mounted () {
-    document.addEventListener('keydown', this.handleDetectShiftKey)
-    document.addEventListener('keyup', this.handleDetectShiftKey)
-  }
+  document.addEventListener('keydown', handleDetectShiftKey)
+  document.addEventListener('keyup', handleDetectShiftKey)
+})
 
-  handleDetectShiftKey (event: KeyboardEvent) {
-    if (event.key === 'Shift') {
-      if (event.type === 'keydown') {
-        this.isShiftKeyPress = true
-      } else if (event.type === 'keyup') {
-        this.isShiftKeyPress = false
-      }
+function handleDetectShiftKey (event: KeyboardEvent) {
+  if (event.key === 'Shift') {
+    if (event.type === 'keydown') {
+      isShiftKeyPress.value = true
+    } else if (event.type === 'keyup') {
+      isShiftKeyPress.value = false
     }
   }
+}
 
-  get filterList () {
-    return this.getGallery()
+const filterList = computed(() => {
+  return getGallery()
+})
+
+const isAllSelected = computed(() => {
+  const values = Object.values(selectedList)
+  if (values.length === 0) {
+    return false
+  } else {
+    return filterList.value.every(item => {
+      return selectedList[item.id!]
+    })
   }
+})
 
-  get isAllSelected () {
-    const values = Object.values(this.choosedList)
-    if (values.length === 0) {
-      return false
-    } else {
-      return this.filterList.every(item => {
-        return this.choosedList[item.id!]
-      })
-    }
-  }
+function getPicBeds (event: IpcRendererEvent, picBeds: IPicBedType[]) {
+  picBed.value = picBeds
+}
 
-  getPicBeds (event: IpcRendererEvent, picBeds: IPicBedType[]) {
-    this.picBed = picBeds
-  }
-
-  getGallery (): ImgInfo[] {
-    if (this.searchText || this.choosedPicBed.length > 0) {
-      return this.images
-        .filter(item => {
-          let isInChoosedPicBed = true
-          let isIncludesSearchText = true
-          if (this.choosedPicBed.length > 0) {
-            isInChoosedPicBed = this.choosedPicBed.some(type => type === item.type)
-          }
-          if (this.searchText) {
-            isIncludesSearchText = item.fileName?.includes(this.searchText) || false
-          }
-          return isIncludesSearchText && isInChoosedPicBed
-        })
-    } else {
-      return this.images
-    }
-  }
-
-  async updateGallery () {
-    this.images = (await this.$$db.get({ orderBy: 'desc' })).data
-  }
-
-  @Watch('filterList')
-  handleFilterListChange () {
-    this.clearChoosedList()
-  }
-
-  handleChooseImage (val: boolean, index: number) {
-    if (val === true) {
-      this.handleBarActive = true
-      if (this.lastChoosed !== -1 && this.isShiftKeyPress) {
-        const min = Math.min(this.lastChoosed, index)
-        const max = Math.max(this.lastChoosed, index)
-        for (let i = min + 1; i < max; i++) {
-          const id = this.filterList[i].id!
-          this.$set(this.choosedList, id, true)
+function getGallery (): IGalleryItem[] {
+  if (searchText.value || selectedPicBed.value.length > 0) {
+    return images.value
+      .filter(item => {
+        let isInSelectedPicBed = true
+        let isIncludesSearchText = true
+        if (selectedPicBed.value.length > 0) {
+          isInSelectedPicBed = selectedPicBed.value.some(type => type === item.type)
         }
+        if (searchText.value) {
+          isIncludesSearchText = item.fileName?.includes(searchText.value) || false
+        }
+        return isIncludesSearchText && isInSelectedPicBed
+      }).map(item => {
+        return {
+          ...item,
+          src: item.imgUrl || '',
+          key: (item.id || `${Date.now()}`),
+          intro: item.fileName || ''
+        }
+      })
+  } else {
+    return images.value.map(item => {
+      return {
+        ...item,
+        src: item.imgUrl || '',
+        key: (item.id || `${Date.now()}`),
+        intro: item.fileName || ''
       }
-      this.lastChoosed = index
-    }
-  }
-
-  clearChoosedList () {
-    this.isShiftKeyPress = false
-    Object.keys(this.choosedList).forEach(key => {
-      this.choosedList[key] = false
     })
-    this.lastChoosed = -1
   }
+}
 
-  zoomImage (index: number) {
-    this.idx = index
-    this.changeZIndexForGallery(true)
-  }
+async function updateGallery () {
+  images.value = (await $$db.get({ orderBy: 'desc' })).data
+}
 
-  changeZIndexForGallery (isOpen: boolean) {
-    if (isOpen) {
-      // @ts-ignore
-      document.querySelector('.main-content.el-row').style.zIndex = 101
-    } else {
-      // @ts-ignore
-      document.querySelector('.main-content.el-row').style.zIndex = 10
+watch(() => filterList, () => {
+  clearSelectedList()
+})
+
+function handleChooseImage (val: CheckboxValueType, index: number) {
+  if (val === true) {
+    handleBarActive.value = true
+    if (lastSelected.value !== -1 && isShiftKeyPress.value) {
+      const min = Math.min(lastSelected.value, index)
+      const max = Math.max(lastSelected.value, index)
+      for (let i = min + 1; i < max; i++) {
+        const id = filterList.value[i].id!
+        selectedList[id] = true
+      }
     }
+    lastSelected.value = index
   }
+}
 
-  handleClose () {
-    this.idx = null
-    this.changeZIndexForGallery(false)
+function clearSelectedList () {
+  isShiftKeyPress.value = false
+  Object.keys(selectedList).forEach(key => {
+    selectedList[key] = false
+  })
+  lastSelected.value = -1
+}
+
+function zoomImage (index: number) {
+  gallerySliderControl.index = index
+  gallerySliderControl.visible = true
+  changeZIndexForGallery(true)
+}
+
+function changeZIndexForGallery (isOpen: boolean) {
+  if (isOpen) {
+    // @ts-ignore
+    document.querySelector('.main-content.el-row').style.zIndex = 101
+  } else {
+    // @ts-ignore
+    document.querySelector('.main-content.el-row').style.zIndex = 10
   }
+}
 
-  async copy (item: ImgInfo) {
-    const copyLink = await ipcRenderer.invoke(PASTE_TEXT, item)
-    const obj = {
-      title: '复制链接成功',
-      body: copyLink,
-      icon: item.url || item.imgUrl
-    }
-    const myNotification = new Notification(obj.title, obj)
-    myNotification.onclick = () => {
-      return true
-    }
+function handleClose () {
+  gallerySliderControl.index = 0
+  gallerySliderControl.visible = false
+  changeZIndexForGallery(false)
+}
+
+async function copy (item: ImgInfo) {
+  const copyLink = await ipcRenderer.invoke(PASTE_TEXT, getRawData(item))
+  const obj = {
+    title: $T('COPY_LINK_SUCCEED'),
+    body: copyLink
+    // sometimes will cause lagging
+    // icon: item.url || item.imgUrl
   }
+  const myNotification = new Notification(obj.title, obj)
+  myNotification.onclick = () => {
+    return true
+  }
+}
 
-  remove (id: string) {
-    this.$confirm('此操作将把该图片移出相册, 是否继续?', '提示', {
-      confirmButtonText: '确定',
-      cancelButtonText: '取消',
+function remove (id?: string) {
+  if (id) {
+    $confirm($T('TIPS_REMOVE_LINK'), $T('TIPS_NOTICE'), {
+      confirmButtonText: $T('CONFIRM'),
+      cancelButtonText: $T('CANCEL'),
       type: 'warning'
     }).then(async () => {
-      const file = await this.$$db.getById(id)
-      await this.$$db.removeById(id)
-      ipcRenderer.send('removeFiles', [file])
+      const file = await $$db.getById(id)
+      await $$db.removeById(id)
+      sendToMain('removeFiles', [file])
       const obj = {
-        title: '操作结果',
-        body: '删除成功'
+        title: $T('OPERATION_SUCCEED'),
+        body: ''
       }
       const myNotification = new Notification(obj.title, obj)
       myNotification.onclick = () => {
         return true
       }
-      this.updateGallery()
+      updateGallery()
     }).catch((e) => {
       console.log(e)
       return true
     })
   }
+}
 
-  openDialog (item: ImgInfo) {
-    this.imgInfo.id = item.id!
-    this.imgInfo.imgUrl = item.imgUrl as string
-    this.dialogVisible = true
+function openDialog (item: ImgInfo) {
+  imgInfo.id = item.id!
+  imgInfo.imgUrl = item.imgUrl as string
+  dialogVisible.value = true
+}
+
+async function confirmModify () {
+  await $$db.updateById(imgInfo.id, {
+    imgUrl: imgInfo.imgUrl
+  })
+  const obj = {
+    title: $T('CHANGE_IMAGE_URL_SUCCEED'),
+    body: imgInfo.imgUrl
+    // icon: this.imgInfo.imgUrl
   }
-
-  async confirmModify () {
-    await this.$$db.updateById(this.imgInfo.id, {
-      imgUrl: this.imgInfo.imgUrl
-    })
-    const obj = {
-      title: '修改图片URL成功',
-      body: this.imgInfo.imgUrl,
-      icon: this.imgInfo.imgUrl
-    }
-    const myNotification = new Notification(obj.title, obj)
-    myNotification.onclick = () => {
-      return true
-    }
-    this.dialogVisible = false
-    this.updateGallery()
+  const myNotification = new Notification(obj.title, obj)
+  myNotification.onclick = () => {
+    return true
   }
+  dialogVisible.value = false
+  updateGallery()
+}
 
-  choosePicBed (type: string) {
-    const idx = this.choosedPicBed.indexOf(type)
-    if (idx !== -1) {
-      this.choosedPicBed.splice(idx, 1)
-    } else {
-      this.choosedPicBed.push(type)
-    }
-  }
+function cleanSearch () {
+  searchText.value = ''
+}
 
-  cleanSearch () {
-    this.searchText = ''
-  }
+function toggleSelectAll () {
+  const result = !isAllSelected.value
+  filterList.value.forEach(item => {
+    selectedList[item.id!] = result
+  })
+}
 
-  isMultiple (obj: IObj) {
-    return Object.values(obj).some(item => item)
-  }
+function selectMore () {
+  sendRPC(IRPCActionType.GET_GALLERY_MENU_LIST, filterList.value.filter(item => {
+    return selectedList[item.id!]
+  }))
+}
 
-  toggleSelectAll () {
-    const result = !this.isAllSelected
-    this.filterList.forEach(item => {
-      this.$set(this.choosedList, item.id!, result)
-    })
-  }
-
-  multiRemove () {
-    // choosedList -> { [id]: true or false }; true means choosed. false means not choosed.
-    const multiRemoveNumber = Object.values(this.choosedList).filter(item => item).length
-    if (multiRemoveNumber) {
-      this.$confirm(`将在相册中移除刚才选中的 ${multiRemoveNumber} 张图片，是否继续？`, '提示', {
-        confirmButtonText: '确定',
-        cancelButtonText: '取消',
-        type: 'warning'
-      }).then(async () => {
-        const files: IResult<ImgInfo>[] = []
-        const imageIDList = Object.keys(this.choosedList)
-        for (let i = 0; i < imageIDList.length; i++) {
-          const key = imageIDList[i]
-          if (this.choosedList[key]) {
-            const file = await this.$$db.getById<ImgInfo>(key)
-            if (file) {
-              files.push(file)
-              await this.$$db.removeById(key)
-            }
-          }
-        }
-        this.clearChoosedList()
-        this.choosedList = {} // 只有删除才能将这个置空
-        const obj = {
-          title: '操作结果',
-          body: '删除成功'
-        }
-        ipcRenderer.send('removeFiles', files)
-        const myNotification = new Notification(obj.title, obj)
-        myNotification.onclick = () => {
-          return true
-        }
-        this.updateGallery()
-      }).catch(() => {
-        return true
-      })
-    }
-  }
-
-  async multiCopy () {
-    if (Object.values(this.choosedList).some(item => item)) {
-      const copyString: string[] = []
-      // choosedList -> { [id]: true or false }; true means choosed. false means not choosed.
-      const imageIDList = Object.keys(this.choosedList)
+function multiRemove () {
+  // selectedList -> { [id]: true or false }; true means selected. false means not selected.
+  const multiRemoveNumber = Object.values(selectedList).filter(item => item).length
+  if (multiRemoveNumber) {
+    $confirm($T('TIPS_WILL_REMOVE_CHOOSED_IMAGES', {
+      m: multiRemoveNumber
+    }), $T('TIPS_NOTICE'), {
+      confirmButtonText: $T('CONFIRM'),
+      cancelButtonText: $T('CANCEL'),
+      type: 'warning'
+    }).then(async () => {
+      const files: IResult<ImgInfo>[] = []
+      const imageIDList = Object.keys(selectedList)
       for (let i = 0; i < imageIDList.length; i++) {
         const key = imageIDList[i]
-        if (this.choosedList[key]) {
-          const item = await this.$$db.getById<ImgInfo>(key)
-          if (item) {
-            const txt = await ipcRenderer.invoke(PASTE_TEXT, item)
-            copyString.push(txt)
-            this.choosedList[key] = false
+        if (selectedList[key]) {
+          const file = await $$db.getById<ImgInfo>(key)
+          if (file) {
+            files.push(file)
+            await $$db.removeById(key)
           }
         }
       }
+      clearSelectedList()
+      // TODO: check this
+      // selectedList = {} // 只有删除才能将这个置空
       const obj = {
-        title: '批量复制链接成功',
-        body: copyString.join('\n')
+        title: $T('OPERATION_SUCCEED'),
+        body: ''
       }
+      sendToMain('removeFiles', files)
       const myNotification = new Notification(obj.title, obj)
-      clipboard.writeText(copyString.join('\n'))
       myNotification.onclick = () => {
         return true
       }
+      updateGallery()
+    }).catch(() => {
+      return true
+    })
+  }
+}
+
+async function multiCopy () {
+  if (Object.values(selectedList).some(item => item)) {
+    const copyString: string[] = []
+    // selectedList -> { [id]: true or false }; true means selected. false means not selected.
+    const imageIDList = Object.keys(selectedList)
+    for (let i = 0; i < imageIDList.length; i++) {
+      const key = imageIDList[i]
+      if (selectedList[key]) {
+        const item = await $$db.getById<ImgInfo>(key)
+        if (item) {
+          const txt = await ipcRenderer.invoke(PASTE_TEXT, getRawData(item))
+          copyString.push(txt)
+          selectedList[key] = false
+        }
+      }
+    }
+    const obj = {
+      title: $T('BATCH_COPY_LINK_SUCCEED'),
+      body: copyString.join('\n')
+    }
+    const myNotification = new Notification(obj.title, obj)
+    clipboard.writeText(copyString.join('\n'))
+    myNotification.onclick = () => {
+      return true
     }
   }
+}
 
-  toggleHandleBar () {
-    this.handleBarActive = !this.handleBarActive
-  }
+function toggleHandleBar () {
+  handleBarActive.value = !handleBarActive.value
+}
 
-  // getPasteStyle () {
-  //   this.pasteStyle = this.$db.get('settings.pasteStyle') || 'markdown'
-  // }
-  async handlePasteStyleChange (val: string) {
-    this.saveConfig('settings.pasteStyle', val)
-    this.pasteStyle = val
-  }
+async function handlePasteStyleChange (val: string) {
+  saveConfig('settings.pasteStyle', val)
+  pasteStyle.value = val
+}
 
-  beforeDestroy () {
-    ipcRenderer.removeAllListeners('updateGallery')
-    ipcRenderer.removeListener('getPicBeds', this.getPicBeds)
-  }
+onBeforeUnmount(() => {
+  ipcRenderer.removeAllListeners('updateGallery')
+  ipcRenderer.removeListener(GET_PICBEDS, getPicBeds)
+})
+
+onActivated(async () => {
+  pasteStyle.value = (await getConfig('settings.pasteStyle')) || 'markdown'
+})
+
+</script>
+<script lang="ts">
+export default {
+  name: 'GalleryPage'
 }
 </script>
 <style lang='stylus'>
+.PhotoSlider
+  &__BannerIcon
+    &:nth-child(1)
+      display none
+  &__Counter
+    margin-top 20px
 .view-title
   color #eee
   font-size 20px
@@ -466,36 +562,8 @@ export default class extends Vue {
       transform: rotate(180deg)
 #gallery-view
   height 100%
-.item-base
-  background #2E2E2E
-  text-align center
-  padding 5px 0
-  cursor pointer
-  font-size 13px
-  transition all .2s ease-in-out
-  height: 28px
-  box-sizing: border-box
-  &.copy
-    cursor not-allowed
-    background #49B1F5
-    &.active
-      cursor pointer
-      background #1B9EF3
-      color #fff
-  &.delete
-    cursor not-allowed
-    background #F47466
-    &.active
-      cursor pointer
-      background #F15140
-      color #fff
-  &.all-pick
-    cursor not-allowed
-    background #69C282
-    &.active
-      cursor pointer
-      background #44B363
-      color #fff
+  .cursor-pointer
+    cursor pointer
 #gallery-view
   position relative
   .round
@@ -516,17 +584,18 @@ export default class extends Vue {
       height: 287px
       top: 113px
     &__img
-      height 150px
+      // height 150px
       position relative
-      margin-bottom 16px
+      margin-bottom 8px
     &__item
       width 100%
       height 120px
       transition all .2s ease-in-out
       cursor pointer
-      margin-bottom 8px
+      margin-bottom 4px
       overflow hidden
       display flex
+      margin-bottom 6px
       &-fake
         position absolute
         top 0
@@ -538,24 +607,34 @@ export default class extends Vue {
         transform scale(1.1)
       &-img
         width 100%
-        object-fit fill
+        object-fit contain
     &__tool-panel
       color #ddd
+      margin-bottom 4px
+      display flex
+      .el-checkbox
+        height 16px
       i
         cursor pointer
         transition all .2s ease-in-out
-        &.el-icon-document
+        margin-right 4px
+        &.document
           &:hover
             color #49B1F5
-        &.el-icon-edit-outline
+        &.edit
           &:hover
             color #69C282
-        &.el-icon-delete
+        &.delete
           &:hover
             color #F15140
+    &__file-name
+      overflow hidden
+      text-overflow ellipsis
+      white-space nowrap
+      color #ddd
+      font-size 14px
+      margin-bottom 4px
   .handle-bar
     color #ddd
     margin-bottom 10px
-  .el-input__inner
-    border-radius 14px
 </style>
