@@ -2,37 +2,36 @@ import {
   Notification,
   BrowserWindow,
   ipcMain,
-  WebContents
+  WebContents,
+  clipboard
 } from 'electron'
 import dayjs from 'dayjs'
 import picgo from '@core/picgo'
 import db from '~/main/apis/core/datastore'
 import windowManager from 'apis/app/window/windowManager'
-import { IWindowList } from 'apis/app/window/constants'
+import { IWindowList } from '#/types/enum'
 import util from 'util'
-import { IPicGo } from 'picgo/dist/src/types'
-import { showNotification, calcDurationRange } from '~/main/utils/common'
-import { TALKING_DATA_EVENT } from '~/universal/events/constants'
+import { IPicGo } from 'picgo'
+import { showNotification, calcDurationRange, getClipboardFilePath } from '~/main/utils/common'
+import { GET_RENAME_FILE_NAME, RENAME_FILE_NAME, TALKING_DATA_EVENT } from '~/universal/events/constants'
 import logger from '@core/picgo/logger'
-
-const waitForShow = (webcontent: WebContents) => {
-  return new Promise<void>((resolve) => {
-    webcontent.on('did-finish-load', () => {
-      resolve()
-    })
-  })
-}
+import { T } from '~/main/i18n'
+import fse from 'fs-extra'
+import path from 'path'
+import { privacyManager } from '~/main/utils/privacyManager'
+import writeFile from 'write-file-atomic'
+import { CLIPBOARD_IMAGE_FOLDER } from '~/universal/utils/static'
 
 const waitForRename = (window: BrowserWindow, id: number): Promise<string|null> => {
   return new Promise((resolve) => {
     const windowId = window.id
-    ipcMain.once(`rename${id}`, (evt: Event, newName: string) => {
+    ipcMain.once(`${RENAME_FILE_NAME}${id}`, (evt: Event, newName: string) => {
       resolve(newName)
       window.close()
     })
     window.on('close', () => {
       resolve(null)
-      ipcMain.removeAllListeners(`rename${id}`)
+      ipcMain.removeAllListeners(`${RENAME_FILE_NAME}${id}`)
       windowManager.deleteById(windowId)
     })
   })
@@ -60,19 +59,19 @@ class Uploader {
   }
 
   init () {
-    picgo.on('notification', message => {
+    picgo.on('notification', (message: Electron.NotificationConstructorOptions | undefined) => {
       const notification = new Notification(message)
       notification.show()
     })
 
-    picgo.on('uploadProgress', progress => {
+    picgo.on('uploadProgress', (progress: any) => {
       this.webContents?.send('uploadProgress', progress)
     })
     picgo.on('beforeTransform', () => {
       if (db.get('settings.uploadNotification')) {
         const notification = new Notification({
-          title: '上传进度',
-          body: '正在上传'
+          title: T('UPLOAD_PROGRESS'),
+          body: T('UPLOADING')
         })
         notification.show()
       }
@@ -86,14 +85,19 @@ class Uploader {
             let name: undefined | string | null
             let fileName: string | undefined
             if (autoRename) {
-              fileName = dayjs().add(index, 'ms').format('YYYYMMDDHHmmSSS') + item.extname
+              fileName = dayjs().add(index, 'ms').format('YYYYMMDDHHmmssSSS') + item.extname
             } else {
               fileName = item.fileName
             }
             if (rename) {
               const window = windowManager.create(IWindowList.RENAME_WINDOW)!
-              await waitForShow(window.webContents)
-              window.webContents.send('rename', fileName, window.webContents.id)
+              logger.info('wait for rename window ready...')
+              ipcMain.on(GET_RENAME_FILE_NAME, (evt) => {
+                if (evt.sender.id === window.webContents.id) {
+                  logger.info('rename window ready, wait for rename...')
+                  window.webContents.send(RENAME_FILE_NAME, fileName, item.fileName, window.webContents.id)
+                }
+              })
               name = await waitForRename(window, window.webContents.id)
             }
             item.fileName = name || fileName
@@ -108,8 +112,43 @@ class Uploader {
     return this
   }
 
+  /**
+   * use electron's clipboard image to upload
+   */
+  async uploadWithBuildInClipboard (): Promise<ImgInfo[]|false> {
+    let filePath = ''
+    try {
+      const imgPath = getClipboardFilePath()
+      if (!imgPath) {
+        const nativeImage = clipboard.readImage()
+        if (nativeImage.isEmpty()) {
+          return false
+        }
+        const buffer = nativeImage.toPNG()
+        const baseDir = picgo.baseDir
+        const fileName = `${dayjs().format('YYYYMMDDHHmmssSSS')}.png`
+        filePath = path.join(baseDir, CLIPBOARD_IMAGE_FOLDER, fileName)
+        await writeFile(filePath, buffer)
+        return await this.upload([filePath])
+      } else {
+        return await this.upload([imgPath])
+      }
+    } catch (e: any) {
+      logger.error(e)
+      return false
+    } finally {
+      if (filePath) {
+        fse.unlink(filePath)
+      }
+    }
+  }
+
   async upload (img?: IUploadOption): Promise<ImgInfo[]|false> {
     try {
+      const privacyCheckRes = await privacyManager.check()
+      if (!privacyCheckRes) {
+        throw Error(T('PRIVACY_TIPS'))
+      }
       const startTime = Date.now()
       const output = await picgo.upload(img)
       if (Array.isArray(output) && output.some((item: ImgInfo) => item.imgUrl)) {
@@ -129,12 +168,14 @@ class Uploader {
       logger.error(e)
       setTimeout(() => {
         showNotification({
-          title: '上传失败',
+          title: T('UPLOAD_FAILED'),
           body: util.format(e.stack),
           clickToCopy: true
         })
       }, 500)
       return false
+    } finally {
+      ipcMain.removeAllListeners(GET_RENAME_FILE_NAME)
     }
   }
 }
