@@ -1,8 +1,10 @@
 import fs from 'fs-extra'
+import writeFile from 'write-file-atomic'
 import path from 'path'
 import { app as APP } from 'electron'
 import { getLogger } from '@core/utils/localLogger'
 import dayjs from 'dayjs'
+import { T } from '~/main/i18n'
 const STORE_PATH = APP.getPath('userData')
 const configFilePath = path.join(STORE_PATH, 'data.json')
 const configFileBackupPath = path.join(STORE_PATH, 'data.bak.json')
@@ -11,8 +13,8 @@ let _configFilePath = ''
 let hasCheckPath = false
 
 const errorMsg = {
-  broken: 'PicGo 配置文件损坏，已经恢复为默认配置',
-  brokenButBackup: 'PicGo 配置文件损坏，已经恢复为备份配置'
+  broken: T('TIPS_PICGO_CONFIG_FILE_BROKEN_WITH_DEFAULT'),
+  brokenButBackup: T('TIPS_PICGO_CONFIG_FILE_BROKEN_WITH_BACKUP')
 }
 
 /** ensure notification list */
@@ -36,7 +38,7 @@ function dbChecker () {
     }
     let configFile: string = '{}'
     const optionsTpl = {
-      title: '注意',
+      title: T('TIPS_NOTICE'),
       body: ''
     }
     // config save bak
@@ -49,9 +51,11 @@ function dbChecker () {
         try {
           configFile = fs.readFileSync(configFileBackupPath, { encoding: 'utf-8' })
           JSON.parse(configFile)
-          fs.writeFileSync(configFilePath, configFile, { encoding: 'utf-8' })
+          writeFile.sync(configFilePath, configFile, { encoding: 'utf-8' })
           const stats = fs.statSync(configFileBackupPath)
-          optionsTpl.body = `${errorMsg.brokenButBackup}\n备份文件版本：${dayjs(stats.mtime).format('YYYY-MM-DD HH:mm:ss')}`
+          optionsTpl.body = `${errorMsg.brokenButBackup}\n${T('TIPS_PICGO_BACKUP_FILE_VERSION', {
+            v: dayjs(stats.mtime).format('YYYY-MM-DD HH:mm:ss')
+          })}`
           global.notificationList?.push(optionsTpl)
           return
         } catch (e) {
@@ -64,7 +68,7 @@ function dbChecker () {
       global.notificationList?.push(optionsTpl)
       return
     }
-    fs.writeFileSync(configFileBackupPath, configFile, { encoding: 'utf-8' })
+    writeFile.sync(configFileBackupPath, configFile, { encoding: 'utf-8' })
   }
 }
 
@@ -94,18 +98,17 @@ function dbPathChecker (): string {
     }
     return _configFilePath
   } catch (e) {
-    const picgoLogPath = path.join(defaultConfigPath, 'picgo.log')
+    const picgoLogPath = path.join(STORE_PATH, 'picgo-gui-local.log')
     const logger = getLogger(picgoLogPath)
     if (!hasCheckPath) {
       const optionsTpl = {
-        title: '注意',
-        body: '自定义文件解析出错，请检查路径内容是否正确'
+        title: T('TIPS_NOTICE'),
+        body: T('TIPS_CUSTOM_CONFIG_FILE_PATH_ERROR')
       }
       global.notificationList?.push(optionsTpl)
       hasCheckPath = true
     }
     logger('error', e)
-    console.error(e)
     _configFilePath = defaultConfigPath
     return _configFilePath
   }
