@@ -4,7 +4,6 @@ import {
   Notification,
   ipcMain
 } from 'electron'
-import path from 'path'
 import db, { GalleryDB } from 'apis/core/datastore'
 import { dbPathChecker, defaultConfigPath, getGalleryDBPath } from 'apis/core/datastore/dbChecker'
 import uploader from 'apis/app/uploader'
@@ -18,15 +17,17 @@ import {
   SHOW_INPUT_BOX
 } from '~/universal/events/constants'
 import { DBStore } from '@picgo/store'
-type PromiseResType<T> = T extends Promise<infer R> ? R : T
+import { T } from '~/main/i18n'
+import { IRPCActionType } from '~/universal/types/enum'
 
 // Cross-process support may be required in the future
 class GuiApi implements IGuiApi {
+  // eslint-disable-next-line no-use-before-define
   private static instance: GuiApi
   private windowId: number = -1
   private settingWindowId: number = -1
   private constructor () {
-    console.log('init guiapi')
+    console.log('init gui api')
   }
 
   public static getInstance (): GuiApi {
@@ -50,7 +51,7 @@ class GuiApi implements IGuiApi {
     })
   }
 
-  private getWebcontentsByWindowId (id: number) {
+  private getWebContentsByWindowId (id: number) {
     return BrowserWindow.fromId(id)?.webContents
   }
 
@@ -59,7 +60,7 @@ class GuiApi implements IGuiApi {
     placeholder: ''
   }) {
     await this.showSettingWindow()
-    this.getWebcontentsByWindowId(this.settingWindowId)?.send(SHOW_INPUT_BOX, options)
+    this.getWebContentsByWindowId(this.settingWindowId)?.send(SHOW_INPUT_BOX, options)
     return new Promise<string>((resolve) => {
       ipcMain.once(SHOW_INPUT_BOX, (event: Event, value: string) => {
         resolve(value)
@@ -70,12 +71,12 @@ class GuiApi implements IGuiApi {
   async showFileExplorer (options: IShowFileExplorerOption = {}) {
     this.windowId = await getWindowId()
     const res = await dialog.showOpenDialog(BrowserWindow.fromId(this.windowId)!, options)
-    return res.filePaths?.[0]
+    return res.filePaths || []
   }
 
   async upload (input: IUploadOption) {
     this.windowId = await getWindowId()
-    const webContents = this.getWebcontentsByWindowId(this.windowId)
+    const webContents = this.getWebContentsByWindowId(this.windowId)
     const imgs = await uploader.setWebContents(webContents!).upload(input)
     if (imgs !== false) {
       const pasteStyle = db.get('settings.pasteStyle') || 'markdown'
@@ -83,9 +84,9 @@ class GuiApi implements IGuiApi {
       for (let i = 0; i < imgs.length; i++) {
         pasteText.push(pasteTemplate(pasteStyle, imgs[i], db.get('settings.customLink')))
         const notification = new Notification({
-          title: '上传成功',
-          body: imgs[i].imgUrl as string,
-          icon: imgs[i].imgUrl
+          title: T('UPLOAD_SUCCEED'),
+          body: imgs[i].imgUrl as string
+          // icon: imgs[i].imgUrl
         })
         setTimeout(() => {
           notification.show()
@@ -94,7 +95,7 @@ class GuiApi implements IGuiApi {
       }
       handleCopyUrl(pasteText.join('\n'))
       webContents?.send('uploadFiles', imgs)
-      webContents?.send('updateGallery')
+      webContents?.send(IRPCActionType.UPDATE_GALLERY)
       return imgs
     }
     return []
@@ -132,6 +133,20 @@ class GuiApi implements IGuiApi {
   }
 
   /**
+   * v2.4.0+
+   * @param options
+   */
+  async showConfigDialog<T extends IStringKeyMap> (options: IPicGoPluginShowConfigDialogOption) {
+    await this.showSettingWindow()
+    this.getWebContentsByWindowId(this.settingWindowId)?.send(IRPCActionType.OPEN_CONFIG_DIALOG, options)
+    return new Promise<T | false>((resolve) => {
+      ipcMain.once(IRPCActionType.OPEN_CONFIG_DIALOG, (event: Event, value: T | false) => {
+        resolve(value)
+      })
+    })
+  }
+
+  /**
    * get picgo config/data path
    */
   async getConfigPath () {
@@ -147,14 +162,35 @@ class GuiApi implements IGuiApi {
   get galleryDB (): DBStore {
     return new Proxy<DBStore>(GalleryDB.getInstance(), {
       get (target, prop: keyof DBStore) {
+        if (prop === 'overwrite') {
+          return new Proxy(GalleryDB.getInstance().overwrite, {
+            apply (target, ctx, args) {
+              return new Promise((resolve) => {
+                const guiApi = GuiApi.getInstance()
+                guiApi.showMessageBox({
+                  title: T('TIPS_WARNING'),
+                  message: T('TIPS_PLUGIN_REMOVE_GALLERY_ITEM'),
+                  type: 'info',
+                  buttons: ['Yes', 'No']
+                }).then(res => {
+                  if (res.result === 0) {
+                    resolve(Reflect.apply(target, ctx, args))
+                  } else {
+                    resolve(undefined)
+                  }
+                })
+              })
+            }
+          })
+        }
         if (prop === 'removeById') {
           return new Proxy(GalleryDB.getInstance().removeById, {
             apply (target, ctx, args) {
               return new Promise((resolve) => {
                 const guiApi = GuiApi.getInstance()
                 guiApi.showMessageBox({
-                  title: '警告',
-                  message: '有插件正在试图删除一些相册图片，是否继续',
+                  title: T('TIPS_WARNING'),
+                  message: T('TIPS_PLUGIN_REMOVE_GALLERY_ITEM'),
                   type: 'info',
                   buttons: ['Yes', 'No']
                 }).then(res => {
