@@ -7,7 +7,7 @@ import {
   BrowserWindow
 } from 'electron'
 import windowManager from 'apis/app/window/windowManager'
-import { IWindowList } from 'apis/app/window/constants'
+import { IRPCActionType, IWindowList } from '#/types/enum'
 import uploader from 'apis/app/uploader'
 import pasteTemplate from '~/main/utils/pasteTemplate'
 import db, { GalleryDB } from '~/main/apis/core/datastore'
@@ -25,8 +25,9 @@ import {
   SHOW_UPLOAD_PAGE_MENU,
   OPEN_USER_STORE_FILE,
   OPEN_URL,
-  RELOAD_APP,
-  SHOW_PLUGIN_PAGE_MENU
+  SHOW_PLUGIN_PAGE_MENU,
+  SET_MINI_WINDOW_POS,
+  GET_PICBEDS
 } from '#/events/constants'
 import {
   uploadClipboardFiles,
@@ -34,8 +35,9 @@ import {
 } from '~/main/apis/app/uploader/apis'
 import picgoCoreIPC from './picgoCoreIPC'
 import { handleCopyUrl } from '~/main/utils/common'
-import { buildMainPageMenu, buildMiniPageMenu, buildPluginPageMenu, buildUploadPageMenu } from './remotes/menu'
+import { buildMainPageMenu, buildMiniPageMenu, buildPluginPageMenu, buildPicBedListMenu } from './remotes/menu'
 import path from 'path'
+import { T } from '~/main/i18n'
 
 const STORE_PATH = app.getPath('userData')
 
@@ -45,27 +47,29 @@ export default {
     // from macOS tray
     ipcMain.on('uploadClipboardFiles', async () => {
       const trayWindow = windowManager.get(IWindowList.TRAY_WINDOW)!
-      const img = await uploader.setWebContents(trayWindow.webContents).upload()
+      // macOS use builtin clipboard is OK
+      const img = await uploader.setWebContents(trayWindow.webContents).uploadWithBuildInClipboard()
       if (img !== false) {
         const pasteStyle = db.get('settings.pasteStyle') || 'markdown'
         handleCopyUrl(pasteTemplate(pasteStyle, img[0], db.get('settings.customLink')))
         const notification = new Notification({
-          title: '上传成功',
-          body: img[0].imgUrl!,
+          title: T('UPLOAD_SUCCEED'),
+          body: img[0].imgUrl!
           // icon: file[0]
-          icon: img[0].imgUrl
+          // icon: img[0].imgUrl
         })
         notification.show()
         await GalleryDB.getInstance().insert(img[0])
         trayWindow.webContents.send('clipboardFiles', [])
         if (windowManager.has(IWindowList.SETTING_WINDOW)) {
-          windowManager.get(IWindowList.SETTING_WINDOW)!.webContents.send('updateGallery')
+          windowManager.get(IWindowList.SETTING_WINDOW)!.webContents.send(IRPCActionType.UPDATE_GALLERY)
         }
       }
       trayWindow.webContents.send('uploadFiles')
     })
 
     ipcMain.on('uploadClipboardFilesFromUploadPage', () => {
+      console.log('handle')
       uploadClipboardFiles()
     })
 
@@ -78,14 +82,14 @@ export default {
       evt.sender.send('updateShortKeyResponse', result)
       if (result) {
         const notification = new Notification({
-          title: '操作成功',
-          body: '你的快捷键已经修改成功'
+          title: T('OPERATION_SUCCEED'),
+          body: T('TIPS_SHORTCUT_MODIFIED_SUCCEED')
         })
         notification.show()
       } else {
         const notification = new Notification({
-          title: '操作失败',
-          body: '快捷键冲突，请重新设置'
+          title: T('OPERATION_FAILED'),
+          body: T('TIPS_SHORTCUT_MODIFIED_CONFLICT')
         })
         notification.show()
       }
@@ -95,14 +99,14 @@ export default {
       const result = shortKeyHandler.bindOrUnbindShortKey(item, from)
       if (result) {
         const notification = new Notification({
-          title: '操作成功',
-          body: '你的快捷键已经修改成功'
+          title: T('OPERATION_SUCCEED'),
+          body: T('TIPS_SHORTCUT_MODIFIED_SUCCEED')
         })
         notification.show()
       } else {
         const notification = new Notification({
-          title: '操作失败',
-          body: '快捷键冲突，请重新设置'
+          title: T('OPERATION_FAILED'),
+          body: T('TIPS_SHORTCUT_MODIFIED_CONFLICT')
         })
         notification.show()
       }
@@ -110,8 +114,8 @@ export default {
 
     ipcMain.on('updateCustomLink', () => {
       const notification = new Notification({
-        title: '操作成功',
-        body: '你的自定义链接格式已经修改成功'
+        title: T('OPERATION_SUCCEED'),
+        body: T('TIPS_CUSTOM_LINK_STYLE_MODIFIED_SUCCEED')
       })
       notification.show()
     })
@@ -132,6 +136,11 @@ export default {
     ipcMain.on('openMiniWindow', () => {
       const miniWindow = windowManager.get(IWindowList.MINI_WINDOW)!
       const settingWindow = windowManager.get(IWindowList.SETTING_WINDOW)!
+
+      if (db.get('settings.miniWindowOnTop')) {
+        miniWindow.setAlwaysOnTop(true)
+      }
+
       miniWindow.show()
       miniWindow.focus()
       settingWindow.hide()
@@ -144,9 +153,9 @@ export default {
       }
     })
 
-    ipcMain.on('getPicBeds', (evt: IpcMainEvent) => {
+    ipcMain.on(GET_PICBEDS, (evt: IpcMainEvent) => {
       const picBeds = getPicBeds()
-      evt.sender.send('getPicBeds', picBeds)
+      evt.sender.send(GET_PICBEDS, picBeds)
       evt.returnValue = picBeds
     })
 
@@ -170,14 +179,14 @@ export default {
     })
     ipcMain.on(SHOW_MAIN_PAGE_MENU, () => {
       const window = windowManager.get(IWindowList.SETTING_WINDOW)!
-      const menu = buildMainPageMenu()
+      const menu = buildMainPageMenu(window)
       menu.popup({
         window
       })
     })
     ipcMain.on(SHOW_UPLOAD_PAGE_MENU, () => {
       const window = windowManager.get(IWindowList.SETTING_WINDOW)!
-      const menu = buildUploadPageMenu()
+      const menu = buildPicBedListMenu()
       menu.popup({
         window
       })
@@ -208,9 +217,9 @@ export default {
     ipcMain.on(OPEN_URL, (evt: IpcMainEvent, url: string) => {
       shell.openExternal(url)
     })
-    ipcMain.on(RELOAD_APP, () => {
-      app.relaunch()
-      app.exit(0)
+    ipcMain.on(SET_MINI_WINDOW_POS, (evt: IpcMainEvent, pos: IMiniWindowPos) => {
+      const window = BrowserWindow.getFocusedWindow()
+      window?.setBounds(pos)
     })
   },
   dispose () {}
