@@ -1,8 +1,10 @@
+import fs from 'fs-extra'
 import db from '~/main/apis/core/datastore'
 import { clipboard, Notification, dialog } from 'electron'
+import { handleUrlEncode } from '~/universal/utils/common'
 
 export const handleCopyUrl = (str: string): void => {
-  if (db.get('settings.autoCopy') !== false) {
+  if (db.get('settings.autoCopyUrl') !== false) {
     clipboard.writeText(str)
   }
 }
@@ -14,16 +16,21 @@ export const handleCopyUrl = (str: string): void => {
 export const showNotification = (options: IPrivateShowNotificationOption = {
   title: '',
   body: '',
-  clickToCopy: false
+  clickToCopy: false,
+  copyContent: '',
+  clickFn: () => {}
 }) => {
   const notification = new Notification({
     title: options.title,
-    body: options.body,
-    icon: options.icon || undefined
+    body: options.body
+    // icon: options.icon || undefined
   })
   const handleClick = () => {
     if (options.clickToCopy) {
-      clipboard.writeText(options.body)
+      clipboard.writeText(options.copyContent || options.body)
+    }
+    if (options.clickFn) {
+      options.clickFn()
     }
   }
   notification.once('click', handleClick)
@@ -68,4 +75,81 @@ export const calcDurationRange = (duration: number) => {
   }
   // max range
   return 100000
+}
+
+/**
+ * macOS public.file-url will get encoded file path,
+ * so we need to decode it
+ */
+export const ensureFilePath = (filePath: string, prefix = 'file://'): string => {
+  filePath = filePath.replace(prefix, '')
+  if (fs.existsSync(filePath)) {
+    return `${prefix}${filePath}`
+  }
+  filePath = decodeURIComponent(filePath)
+  if (fs.existsSync(filePath)) {
+    return `${prefix}${filePath}`
+  }
+  return ''
+}
+
+/**
+ * for builtin clipboard to get image path from clipboard
+ * @returns
+ */
+export const getClipboardFilePath = (): string => {
+  // TODO: linux support
+  const img = clipboard.readImage()
+  if (img.isEmpty()) {
+    if (process.platform === 'win32') {
+      const imgPath = clipboard.readBuffer('FileNameW')?.toString('ucs2')?.replace(RegExp(String.fromCharCode(0), 'g'), '')
+      if (imgPath) {
+        return imgPath
+      }
+    }
+  } else {
+    if (process.platform === 'darwin') {
+      let imgPath = clipboard.read('public.file-url') // will get file://xxx/xxx
+      imgPath = ensureFilePath(imgPath)
+      if (imgPath) {
+        return imgPath.replace('file://', '')
+      }
+    }
+  }
+  return ''
+}
+
+export const handleUrlEncodeWithSetting = (url: string) => {
+  if (db.get('settings.encodeOutputURL') === true) {
+    url = handleUrlEncode(url)
+  }
+  return url
+}
+
+export const replaceHost = (url: string, oldHost: string, newHost: string) => {
+  try {
+    const parsedUrl = new URL(url)
+    if (parsedUrl.host === oldHost) {
+      parsedUrl.host = newHost
+    }
+    return parsedUrl.toString()
+  } catch {
+    return url
+  }
+}
+
+export const getHost = (url: string = '') => {
+  try {
+    const parsedUrl = new URL(url)
+    return parsedUrl.host
+  } catch {
+    return ''
+  }
+}
+
+/**
+ * remove protocol and suffix
+ */
+export const removeProtocolAndSuffix = (url: string = '') => {
+  return url.replace(/^(https?:\/\/)?/, '').replace(/\/$/, '')
 }
