@@ -1,80 +1,43 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import { createRouter, createWebHashHistory } from 'vue-router'
+import * as config from './config'
 
-Vue.use(Router)
-
-export default new Router({
-  mode: 'hash',
+export default createRouter({
+  history: createWebHashHistory(),
   routes: [
     {
       path: '/',
-      name: 'tray-page',
+      name: config.TRAY_PAGE,
       component: () => import(/* webpackChunkName: "tray" */ '@/pages/TrayPage.vue')
     },
     {
       path: '/rename-page',
-      name: 'rename-page',
+      name: config.RENAME_PAGE,
       component: () => import(/* webpackChunkName: "RenamePage" */ '@/pages/RenamePage.vue')
     },
     {
       path: '/mini-page',
-      name: 'mini-page',
+      name: config.MINI_PAGE,
       component: () => import(/* webpackChunkName: "MiniPage" */ '@/pages/MiniPage.vue')
     },
     {
       path: '/main-page',
-      name: 'main-page',
+      name: config.MAIN_PAGE,
       component: () => import(/* webpackChunkName: "SettingPage" */ '@/layouts/Main.vue'),
       children: [
         {
           path: 'upload',
           component: () => import(/* webpackChunkName: "Upload" */ '@/pages/Upload.vue'),
-          name: 'upload'
+          name: config.UPLOAD_PAGE
         },
         {
-          path: 'qiniu',
-          component: () => import(/* webpackChunkName: "Qiniu" */ '@/pages/picbeds/Qiniu.vue'),
-          name: 'qiniu'
-        },
-        {
-          path: 'tcyun',
-          component: () => import(/* webpackChunkName: "TcYun" */ '@/pages/picbeds/TcYun.vue'),
-          name: 'tcyun'
-        },
-        {
-          path: 'upyun',
-          component: () => import(/* webpackChunkName: "UpYun" */ '@/pages/picbeds/UpYun.vue'),
-          name: 'upyun'
-        },
-        {
-          path: 'github',
-          component: () => import(/* webpackChunkName: "GitHub" */ '@/pages/picbeds/GitHub.vue'),
-          name: 'github'
-        },
-        {
-          path: 'smms',
-          component: () => import(/* webpackChunkName: "SMMS" */ '@/pages/picbeds/SMMS.vue'),
-          name: 'smms'
-        },
-        {
-          path: 'aliyun',
-          component: () => import(/* webpackChunkName: "AliYun" */ '@/pages/picbeds/AliYun.vue'),
-          name: 'aliyun'
-        },
-        {
-          path: 'imgur',
-          component: () => import(/* webpackChunkName: "Imgur" */ '@/pages/picbeds/Imgur.vue'),
-          name: 'imgur'
-        },
-        {
-          path: 'others/:type',
-          component: () => import(/* webpackChunkName: "Other" */ '@/pages/picbeds/Others.vue'),
-          name: 'others'
+          path: 'picbeds/:type/:configId?',
+          component: () => import(/* webpackChunkName: "Other" */ '@/pages/picbeds/index.vue'),
+          name: config.PICBEDS_PAGE
         },
         {
           path: 'gallery',
-          component: () => import(/* webpackChunkName: "Gallery" */ '@/pages/Gallery.vue'),
-          name: 'gallery',
+          component: () => import(/* webpackChunkName: "GalleryView" */ '@/pages/Gallery.vue'),
+          name: config.GALLERY_PAGE,
           meta: {
             keepAlive: true
           }
@@ -82,22 +45,32 @@ export default new Router({
         {
           path: 'setting',
           component: () => import(/* webpackChunkName: "setting" */ '@/pages/PicGoSetting.vue'),
-          name: 'setting'
+          name: config.SETTING_PAGE
         },
         {
           path: 'plugin',
           component: () => import(/* webpackChunkName: "Plugin" */ '@/pages/Plugin.vue'),
-          name: 'plugin'
+          name: config.PLUGIN_PAGE
         },
         {
           path: 'shortKey',
           component: () => import(/* webpackChunkName: "ShortkeyPage" */ '@/pages/ShortKey.vue'),
-          name: 'shortKey'
+          name: config.SHORTKEY_PAGE
+        },
+        {
+          path: 'uploader-config-page/:type',
+          component: () => import(/* webpackChunkName: "Other" */ '@/pages/UploaderConfigPage.vue'),
+          name: config.UPLOADER_CONFIG_PAGE
         }
       ]
     },
     {
-      path: '*',
+      path: '/toolbox-page',
+      name: config.TOOLBOX_CONFIG_PAGE,
+      component: () => import(/* webpackChunkName: "ToolboxPage" */ '@/pages/Toolbox.vue')
+    },
+    {
+      path: '/:pathMatch(.*)*',
       redirect: '/'
     }
   ]
