@@ -1,15 +1,16 @@
-import { IObject, IResult, IGetResult, IFilter } from '@picgo/store/dist/types'
-import { ipcRenderer, IpcRendererEvent } from 'electron'
-import { uuid } from 'uuidv4'
 import {
+  PICGO_GET_BY_ID_DB,
   PICGO_GET_DB,
   PICGO_INSERT_DB,
   PICGO_INSERT_MANY_DB,
-  PICGO_UPDATE_BY_ID_DB,
-  PICGO_GET_BY_ID_DB,
-  PICGO_REMOVE_BY_ID_DB
+  PICGO_REMOVE_BY_ID_DB,
+  PICGO_UPDATE_BY_ID_DB
 } from '#/events/constants'
 import { IGalleryDB } from '#/types/extra-vue'
+import { IFilter, IGetResult, IObject, IResult } from '@picgo/store/dist/types'
+import { IpcRendererEvent, ipcRenderer } from 'electron'
+import { v4 as uuid } from 'uuid'
+import { getRawData } from './common'
 export class GalleryDB implements IGalleryDB {
   async get<T> (filter?: IFilter): Promise<IGetResult<T>> {
     const res = await this.msgHandler<IGetResult<T>>(PICGO_GET_DB, filter)
@@ -50,8 +51,9 @@ export class GalleryDB implements IGalleryDB {
           ipcRenderer.removeListener(method, callback)
         }
       }
+      const data = getRawData(args)
       ipcRenderer.on(method, callback)
-      ipcRenderer.send(method, ...args, callbackId)
+      ipcRenderer.send(method, ...data, callbackId)
     })
   }
 }
