@@ -22,8 +22,22 @@ export const MINIMIZE_WINDOW = 'MINIMIZE_WINDOW'
 export const CLOSE_WINDOW = 'CLOSE_WINDOW'
 export const OPEN_USER_STORE_FILE = 'OPEN_USER_STORE_FILE'
 export const OPEN_URL = 'OPEN_URL'
-export const RELOAD_APP = 'RELOAD_APP'
 export const PICGO_CONFIG_PLUGIN = 'PICGO_CONFIG_PLUGIN'
 export const PICGO_HANDLE_PLUGIN_ING = 'PICGO_HANDLE_PLUGIN_ING'
+export const PICGO_HANDLE_PLUGIN_DONE = 'PICGO_HANDLE_PLUGIN_DONE'
 export const PICGO_TOGGLE_PLUGIN = 'PICGO_TOGGLE_PLUGIN'
 export const PASTE_TEXT = 'PASTE_TEXT'
+export const SET_MINI_WINDOW_POS = 'SET_MINI_WINDOW_POS'
+export const RENAME_FILE_NAME = 'RENAME_FILE_NAME'
+export const GET_RENAME_FILE_NAME = 'GET_RENAME_FILE_NAME'
+export const SHOW_MAIN_PAGE_QRCODE = 'SHOW_MAIN_PAGE_QRCODE'
+export const SHOW_MAIN_PAGE_DONATION = 'SHOW_MAIN_PAGE_DONATION'
+export const FORCE_UPDATE = 'FORCE_UPDATE'
+export const OPEN_WINDOW = 'OPEN_WINDOW'
+export const GET_PICBEDS = 'GET_PICBEDS'
+export const RPC_ACTIONS = 'RPC_ACTIONS'
+export const GET_PICBED_CONFIG = 'GET_PICBED_CONFIG'
+// i18n
+export const GET_CURRENT_LANGUAGE = 'GET_CURRENT_LANGUAGE'
+export const GET_LANGUAGE_LIST = 'GET_LANGUAGE_LIST'
+export const SET_CURRENT_LANGUAGE = 'SET_CURRENT_LANGUAGE'
