@@ -1,9 +1,5 @@
 declare module '*.vue' {
-  import Vue from 'vue'
-  export default Vue
-}
-// third-party
-declare module 'fix-path' {
-  function fixPath(): void
-  export default fixPath
+  import { DefineComponent } from 'vue'
+  const component: DefineComponent<{}, {}, any>
+  export default component
 }
